requirements.txt
requests
django
django-admin-data-from-external-service
python-decouple
